Temporarily ignore function calls and move local variable stuff `GoToDefLocalVariableDeclaration`
